requirements.txt
streamlit>=1.32.0
Pillow>=10.0.0
numpy>=1.24.0
supabase>=2.3.0
qrcode[pil]>=7.4.2
requests>=2.31.0
